### 2021-03-04 #3084 @whedon
[PRE REVIEW]: %PT_GSDesign: A SAS Macro for Group Sequential Designs with Time-to-event Data using the Concept of Proportional Time
**Submitting author:** @thewan05 (<a href="http://orcid.org/0000-0002-9306-2619">Nadeesha Thewarapperuma</a>)
**Repository:** <a href="https://github.com/thewan05/GSD_SAS_Macro" target ="_blank">https://github.com/thewan05/GSD_SAS_Macro</a>
**Version:** v1.0.0
