+++ paper.bib
@@ -52,7 +52,8 @@
   journal   = {Statistical Science},
   year      = {1990},
   volume     = {5},
-  pages = {299-317}
+  pages = {299-317},
+  doi = {10.1214/ss/1177012099}
 }
 
 @book{jennison2,
@@ -60,7 +61,8 @@
   year = {2000},
   title = {Group Sequential Methods with Applications to Clinical Trials},
   publisher = {Chapman \& Hall/CRC},
-  address = {Boca Raton}
+  address = {Boca Raton},
+  doi = {10.1201/9780367805326}
 }
 
 @ARTICLE{lan,
@@ -185,5 +187,6 @@
   journal   = {Statistics in Medicine},
   year      = {1999},
   volume     = {18},
-  pages = {2271-2286}
+  pages = {2271-2286},
+  doi = {10.1002/(sici)1097-0258(19990915/30)18:17/18<2271::aid-sim254>3.0.co;2-z}
 }
