### 2021-03-24 #3084 @drvinceknight
@thewan05 have you been able to take a look at this [list of people](https://bit.ly/joss-reviewers) to see if there is anyone you would suggest as a reviewer?

No problem if you do not see anyone but it's a helpful place to start.
