### 2021-07-10 #3084 @thewan05
@drvinceknight My advisor and I are from the Biostatistics department at U of Kansas Medical Center. There is a faculty member in our department, Dr. John Keighley, who has over 20 years of experience working in SAS and is active in the SAS community. If it's acceptable, Dr. Keighley can review the paper. However, if being from the same department is a conflict of interest, he may be able to suggest other reviewers. The reviewers will be SAS certified but may not have a PhD. Is that still fine? Thank you and let me know what you think.
