### 2021-07-26 #3084 @drvinceknight
Apologies for the late reply (I've been away).

> The reviewers will be SAS certified but may not have a PhD. Is that still fine? Thank you and let me know what you think.

That would be fine, having a PhD is not at all a requirement. If in the first instance Dr Keighley could suggest some reviewers that would be appreciated.
