### 2021-10-04 #3084 @thewan05
@drvinceknight Thank you for letting me know. I reached out through the SAS community forums, and one person agreed to review the paper. She is Lauren Rackley and her email address is lauren.g.rackley@gmail.com
