### 2021-10-11 #3813 @whedon
[REVIEW]: %PT_GSDesign: A SAS Macro for Group Sequential Designs with Time-to-event Data using the Concept of Proportional Time
**Submitting author:** @thewan05 (<a href="http://orcid.org/0000-0002-9306-2619">Nadeesha Thewarapperuma</a>)
**Repository:** https://github.com/thewan05/GSD_SAS_Macro
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
