### 2021-12-07 #3813 @thewan05
@drvinceknight Dr. Knight, I'm not sure what the review process is like at JOSS, but can you provide an update? Thank you.
