### 2022-02-02 #3813 @drvinceknight
> @drvinceknight Dr. Knight, I'm not sure what the review process is like at JOSS, but can you provide an update? Thank you.

If I don't hear back @thewan05 in a week or so I'll email them outside of GitHub.
