### 2022-05-25 #3813 @thewan05
@drvinceknight Thank you for your help in reviewing our submission. It looks like it is very hard to find SAS reviewers. We would like to withdraw this paper from JOSS. Kindly guide us through this process.
